# api/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Ship API",
    "version": "1.0.0",
    "description": "Ship is a project and sprint management platform with real-time collaboration.\n\n## Authentication\n\nAll endpoints (except /auth/login and /health) require authentication via:\n- **Session Cookie**: Automatically set after login (15-minute timeout, 12-hour absolute)\n- **Bearer Token**: API tokens from Settings > API Tokens\n\n## Core Concepts\n\n- **Documents**: Everything in Ship is a document (wikis, issues, projects, sprints, etc.)\n- **Document Type**: Each document has a `document_type` that determines its properties\n- **Belongs To**: Documents can be associated via `belongs_to` array (programs, projects, sprints, parent issues)\n\n## WebSocket Collaboration\n\nReal-time editing available at `/collaboration/{docType}:{docId}` using Yjs CRDT protocol.",
    "contact": {
      "name": "Ship Team"
    }
  },
  "servers": [
    {
      "url": "/api",
      "description": "API base path"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    },
    {
      "cookieAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API token authentication. Get your token from Settings > API Tokens."
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session",
        "description": "Session cookie authentication. Automatically set after login."
      }
    },
    "schemas": {
...
      "UploadRequest": {
        "type": "object",
        "properties": {
          "filename": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255,
            "description": "Original filename",
            "example": "screenshot.png"
          },
          "mimeType": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "description": "MIME type of the file",
            "example": "image/png"
          },
          "sizeBytes": {
            "type": "integer",
            "minimum": 0,
            "exclusiveMinimum": true,
            "maximum": 1073741824,
            "description": "File size in bytes (max 1GB)",
            "example": 1024000
          }
        },
        "required": [
          "filename",
          "mimeType",
          "sizeBytes"
        ]
      },
      "UploadResponse": {
        "type": "object",
        "properties": {
          "uploadUrl": {
            "type": "string",
            "format": "uri",
            "description": "Presigned URL to PUT the file (expires in 15 minutes)"
          },
          "fileId": {
            "type": "string",
            "format": "uuid",
            "description": "File ID to use when referencing this file",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "publicUrl": {
            "type": "string",
            "format": "uri",
            "description": "URL where the file will be accessible after upload"
          }
        },
        "required": [
          "uploadUrl",
          "fileId",
          "publicUrl"
        ]
      },
      "FileMetadata": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "UUID identifier",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "filename": {
            "type": "string"
          },
          "mimeType": {
            "type": "string"
          },
          "sizeBytes": {
            "type": "integer"
          },
          "uploadedBy": {
            "type": "string",
            "format": "uuid",
            "description": "UUID identifier",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "documentId": {
            "type": "string",
            "nullable": true,
            "format": "uuid",
            "description": "Document this file is attached to",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "publicUrl": {
            "type": "string",
            "format": "uri"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "ISO 8601 datetime string",
            "example": "2025-01-30T14:30:00.000Z"
          }
        },
        "required": [
          "id",
          "filename",
          "mimeType",
          "sizeBytes",
          "uploadedBy",
          "documentId",
          "publicUrl",
          "createdAt"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/files/upload": {
      "post": {
        "tags": [
          "Files"
        ],
        "summary": "Get presigned upload URL",
        "description": "Request a presigned URL to upload a file. Upload the file via PUT to the returned URL.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Upload URL and file metadata",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or blocked file type",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "blockedExtensions": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/files/{fileId}": {
      "get": {
        "tags": [
          "Files"
        ],
        "summary": "Get file metadata",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "UUID identifier",
              "example": "550e8400-e29b-41d4-a716-446655440000"
            },
            "required": true,
            "description": "UUID identifier",
            "name": "fileId",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "File metadata",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileMetadata"
                }
              }
            }
          },
          "404": {
            "description": "File not found"
          }
        }
      },
      "delete": {
        "tags": [
          "Files"
        ],
        "summary": "Delete file",
        "description": "Delete a file. Only the uploader or an admin can delete.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "UUID identifier",
              "example": "550e8400-e29b-41d4-a716-446655440000"
            },
            "required": true,
            "description": "UUID identifier",
            "name": "fileId",
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "File deleted"
          },
          "403": {
            "description": "Forbidden - not the uploader or admin"
          },
          "404": {
            "description": "File not found"
          }
        }
      }
    },
...
  }
}